+++ CHANGELOG.md
@@ -50,6 +50,7 @@ Many bug fixes to coincide with release of Zapit on eLife.
     this means a limit of about 55 data points per condition. So that's a lot and we will
     probably run out of laser power first. 
   * Stimulus configs loaded via the API now update the GUI. 
+  * BUGFIX: Scanner calibration points are now centered. Previously they were offset. 
   * BUGFIX: Correctly subtract background image when running getLaserPosAccuracy. 
     This uses a different threshold. Also correct bug that was failing to feed that image to
     this function.